stworzone logowanie i wylogowywanie

+++ korepetycje2/src/main/java/org/example/lesson/LessonDTO.java
@@ -8,8 +8,8 @@ public class LessonDTO {
     @JsonFormat(pattern = "yyyy-MM-dd'T'HH:mm:ss")
     private LocalDateTime data;
     private double prize;
-    private long studentId;
-    private long teacherId;
+    private Long studentId;
+    private Long teacherId;
 
     public LocalDateTime getData() { return data; }
     public void setData(LocalDateTime data) { this.data = data; }
@@ -18,8 +18,8 @@ public class LessonDTO {
     public void setPrize(double prize) { this.prize = prize; }
 
     public long getStudentId() { return studentId; }
-    public void setStudentId(int studentId) { this.studentId = studentId; }
+    public void setStudentId(Long studentId) { this.studentId = studentId; }
 
     public long getTeacherId() { return teacherId; }
-    public void setTeacherId(int teacherId) { this.teacherId = teacherId; }
+    public void setTeacherId(Long teacherId) { this.teacherId = teacherId; }
 }
